# zentra/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateStockMovementRequest:
    properties:
      item_id:
        example: 1
        type: integer
      movement_type:
        example: in
        type: string
      notes:
        example: Initial stock
        type: string
      quantity:
        example: 100
        type: integer
      reference_id:
        example: 1
        type: integer
      reference_type:
        example: purchase
        type: string
    required:
    - item_id
    - movement_type
    - quantity
    type: object
...
  handlers.ErrorResponse:
    description: Standard error response model
    properties:
      error:
        example: Invalid request parameters
        type: string
    type: object
...
  handlers.ItemResponse:
    properties:
      code:
        example: ITM001
        type: string
      created_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      created_by:
        example: admin
        type: string
      description:
        example: Test item description
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      max_stock:
        example: 100
        type: integer
      min_stock:
        example: 10
        type: integer
      name:
        example: Test Item
        type: string
      tenant_id:
        example: 1
        type: integer
      unit:
        example: PCS
        type: string
      updated_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      updated_by:
        example: admin
        type: string
    type: object
...
  handlers.StockMovementResponse:
    properties:
      created_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      created_by:
        example: admin
        type: string
      id:
        example: 1
        type: integer
      item:
        $ref: '#/definitions/handlers.ItemResponse'
      movement_type:
        example: in
        type: string
      notes:
        example: Initial stock
        type: string
      quantity:
        example: 100
        type: integer
      reference_id:
        example: 1
        type: integer
      reference_type:
        example: purchase
        type: string
      tenant_id:
        example: 1
        type: integer
      updated_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      updated_by:
        example: admin
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Vomo backend API documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vomo API
  version: "1.0"
paths:
...
  /stock-movements:
    get:
      description: Get all stock movements
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.StockMovementResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all stock movements
      tags:
      - StockMovement
    post:
      consumes:
      - application/json
      description: Create a new stock movement with the provided details
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Stock Movement Data
        in: body
        name: movement
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateStockMovementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.StockMovementResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new stock movement
      tags:
      - StockMovement
  /stock-movements/{id}:
    get:
      description: Get stock movement details by ID
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Stock Movement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StockMovementResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Stock movement not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a stock movement by ID
      tags:
      - StockMovement
  /stock-movements/by-item:
    get:
      description: Get all stock movements for a specific item
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Item ID
        in: query
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.StockMovementResponse'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get stock movements by item
      tags:
      - StockMovement
  /stock-movements/by-reference:
    get:
      description: Get all stock movements for a specific reference
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Reference Type
        in: query
        name: reference_type
        required: true
        type: string
      - description: Reference ID
        in: query
        name: reference_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.StockMovementResponse'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get stock movements by reference
      tags:
      - StockMovement
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"